dom/imptests/html/html/dom/documents/dta/doc.gEBN/test_document.getElementsByName-null-undef.html

<!DOCTYPE html>
<title>Calling getElementsByName with null and undefined</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#dom-document-getelementsbyname">
<link rel="help" href="http://dev.w3.org/2006/webapi/WebIDL/#es-DOMString">
<link rel="help" href="http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-262.pdf#page=57">
<link rel="stylesheet" href="/resources/testharness.css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var n = document.createElement("div");
  n.setAttribute("name", "null");
  document.body.appendChild(n);
  var u = document.createElement("div");
  u.setAttribute("name", "undefined");
  document.body.appendChild(u);
  
//  assert_equals(document.getElementsByName(null)[0], n);
  assert_equals(document.getElementsByName(undefined)[0], u);
});
</script>
